feat: add organization time zone and company domain fields to organization details
